abis/linux: make nlink_t 64-bit.
nlink_t was erroneously defined as 32-bit, which caused some of the
later members of struct stat (st_mode, st_gid, st_uid) to unexpectedly
shift.

+++ abis/linux/x86_64/nlink_t.h
@@ -2,7 +2,7 @@
 #ifndef _ABIBITS_NLINK_T_H
 #define _ABIBITS_NLINK_T_H
 
-typedef int nlink_t;
+typedef unsigned long nlink_t;
 
 #endif // _ABIBITS_NLINK_T_H
